+++ Android.mk
@@ -54,6 +54,7 @@ LOCAL_SRC_FILES := \
 	drmproperty.cpp \
 	glworker.cpp \
 	hwcomposer.cpp \
+	hwcutils.cpp \
         platform.cpp \
         platformdrmgeneric.cpp \
 	platformia.cpp \
